Improve bench, add comment
